+++ src/NetStandard/E.Standard.WebMapping.Core.Api/UI/Elements/UIParagraph.cs
@@ -1,0 +1,15 @@
+﻿using Newtonsoft.Json;
+using System;
+
+namespace E.Standard.WebMapping.Core.Api.UI.Elements;
+
+public class UIParagraph : UIElement
+{
+    public UIParagraph(string text)
+        : base("paragraph")
+        => (Text) = (text ?? String.Empty);
+
+    [JsonProperty(PropertyName = "text", NullValueHandling = NullValueHandling.Ignore)]
+    [System.Text.Json.Serialization.JsonPropertyName("text")]
+    public string Text { get; set; }
+}